+++ ChangeLog.md
@@ -2,7 +2,9 @@
 ## BountyBot Version 2.5.0:
 - Bounty board forums created by `/create-default` now have 👀 as a default reaction
 - BountyBot now provides shortcut links when mentioning its own commands
+- Threads on the bounty board now include a Complete button
 - Bounty descriptions are now optional
+- Adding a description, image, or start and end timestamps each add a 1 XP bonus to the poster's XP on completion
 
 ## BountyBot Version 2.4.0:
 - Ephemeral messages in multi-step processes now clean themselves up